Fitness calculator is taking the parameters from the body form and returns the date to the console we just have to make the data to show on the body form
