+++ client/src/App.js
@@ -16,70 +16,6 @@ import './App.css';
 import { UserProvider } from "./utils/UserContext"
 import axios from 'axios'
 
-const bmi = {
-	method: 'GET',
-	url: 'https://fitness-calculator.p.rapidapi.com/bmi', 
-	params: {age: '25', height: '180', weight: '65'},
-	headers: {
-	  'x-rapidapi-key': '515c74fb86mshcb44e437cf75abcp1b8dc7jsn1ac8f5643c83',
-	  'x-rapidapi-host': 'fitness-calculator.p.rapidapi.com'
-	}
-  };
-  axios.request(bmi).then(function (response) {
-	  console.log(response.data);
-  }).catch(function (error) {
-	  console.error(error);
-  });
-  const idealWeight = {
-	method: 'GET',
-	url: 'https://fitness-calculator.p.rapidapi.com/idealweight',
-	params: {weight: '70', gender: 'male', height: '178'},
-	headers: {
-	  'x-rapidapi-key': '515c74fb86mshcb44e437cf75abcp1b8dc7jsn1ac8f5643c83',
-	  'x-rapidapi-host': 'fitness-calculator.p.rapidapi.com'
-	}
-  };
-  axios.request(idealWeight).then(function (response) {
-	  console.log(response.data);
-  }).catch(function (error) {
-	  console.error(error);
-  });
-  const bodyFat = {
-	method: 'GET',
-	url: 'https://fitness-calculator.p.rapidapi.com/bodyfat',
-	params: {
-	  waist: '96',
-	  gender: 'male',
-	  neck: '50',
-	  heigth: '178',
-	  hip: '92',
-	  age: '25',
-	  weigth: '70'
-	},
-	headers: {
-	  'x-rapidapi-key': '515c74fb86mshcb44e437cf75abcp1b8dc7jsn1ac8f5643c83',
-	  'x-rapidapi-host': 'fitness-calculator.p.rapidapi.com'
-	}
-  };
-  axios.request(bodyFat).then(function (response) {
-	  console.log(response.data);
-  }).catch(function (error) {
-	  console.error(error);
-  });
-  const dailyC = {
-	method: 'GET',
-	url: 'https://fitness-calculator.p.rapidapi.com/dailycalory',
-	params: {heigth: '180', age: '25', gender: 'male', weigth: '70'},
-	headers: {
-	  'x-rapidapi-key': '515c74fb86mshcb44e437cf75abcp1b8dc7jsn1ac8f5643c83',
-	  'x-rapidapi-host': 'fitness-calculator.p.rapidapi.com'
-	}
-  };
-  axios.request(dailyC).then(function (response) {
-	  console.log(response.data);
-  }).catch(function (error) {
-	  console.error(error);
-  });
   const options = {
 	method: 'GET',
 	url: 'https://nutritionix-api.p.rapidapi.com/v1_1/search/bacon',
